Dynamic cast bugfix.

+++ src/benchmark.cpp
@@ -79,7 +79,7 @@ BENCHMARK_F(SmallImageFixture, TraceOrthonormal, 2, 3)
         };
         std::vector<PFunctional> pfunctionals{
                 PFunctional("H1",
-                                new GenericFunctionalWrapper<unsigned int, unsigned int>(PFunctionalHermite),
+                                new GenericFunctionalWrapper<unsigned int, size_t>(PFunctionalHermite),
                                 PFunctional::Type::HERMITE,
                                 1)
         };
